[DONE ] Cow model, added new fuction, utils

+++ src/app/modules/users/user.service.ts
@@ -1,40 +1,103 @@
 import httpStatus from 'http-status';
+import { USER_ENUM } from '../../../enum/common';
+import { PaginationHelpers } from '../../../helpers/pagination.helpers';
 import ApiError from '../../../shared/errors/ApiError';
-import { IUser } from './user.interface';
+import { IGenericResponse } from '../../../shared/interfaces/common.interface';
+import { IPaginationOptions } from '../../../shared/pagination/pagination.interface';
+import { userSearchFields } from './user.constant';
+import { IUser, IUserSearchFilter } from './user.interface';
 import { User } from './user.model';
+import { isUserAvailable } from './user.utils';
 
-//------create a new User service --------------------------------
-const createUser = async (user: IUser): Promise<IUser | null> => {
-  user.budget = user.budget ?? 0;
-  user.income = user.income ?? 0;
-  const savedUser = await User.create(user);
-  return savedUser;
+//------signUp / create a new User service --------------------------------
+const signUp = async (payload: IUser): Promise<IUser | null> => {
+  if (payload.role == USER_ENUM.BUYER) {
+    if (!payload.budget) {
+      throw new ApiError(httpStatus.BAD_REQUEST, 'Budget not specified !');
+    }
+    if (payload.budget < 15000) {
+      throw new ApiError(
+        httpStatus.BAD_REQUEST,
+        'Budget must be at least more than 15000',
+      );
+    }
+  }
+
+  if (payload.role == USER_ENUM.SELLER) payload.income = 0;
+
+  const data = await User.create(payload);
+  return data;
 };
 
 //------Get all user service --------------------------------
-const getAllUsers = async () => {
-  const allUsers = await User.find();
-  return allUsers;
+const getAllUsers = async (
+  paginationOptions: IPaginationOptions,
+  searchFilterFields: IUserSearchFilter,
+): Promise<IGenericResponse<IUser[]> | null> => {
+  // Pagination
+  const { page, limit, skip, sortBy, sortOrder } =
+    PaginationHelpers(paginationOptions);
+
+  // Sort Condition
+  const sortCondition = { [sortBy]: sortOrder };
+
+  const { searchTerm, ...filterData } = searchFilterFields;
+  const andCondition = [];
+
+  // Search Condition
+  if (searchTerm) {
+    andCondition.push({
+      $or: userSearchFields.map(field => ({
+        [field]: { $regex: searchTerm, $options: 'i' },
+      })),
+    });
+  }
+
+  // Filter Fields
+  if (Object.keys(filterData).length) {
+    andCondition.push({
+      $and: Object.entries(filterData).map(([field, value]) => ({
+        [field]: [value],
+      })),
+    });
+  }
+
+  const whereCondition = andCondition.length ? { $and: andCondition } : {};
+
+  const data = await User.find(whereCondition)
+    .sort(sortCondition)
+    .skip(skip)
+    .limit(limit);
+  const total = await User.countDocuments();
+  const meta = { page, limit, total };
+  return { meta, data };
 };
 
 //------get a single user service --------------------------------
 const getSingleUser = async (id: string): Promise<IUser | null> => {
+  if (!(await isUserAvailable(id))) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'User not Found!');
+  }
+
   const singleUser = await User.findById(id);
   return singleUser;
 };
 
 //------ update a User service --------------------------------
 const updateUser = async (
-  user: Partial<IUser>,
-  id: string,
+  _id: string,
+  payload: IUser,
 ): Promise<IUser | null> => {
-  const isExist = await User.findById(id);
-  if (!isExist) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'User not found in the database');
+  if (!(await isUserAvailable(_id))) {
+    throw new ApiError(
+      httpStatus.NOT_FOUND,
+      'User not found in the Database !',
+    );
   }
 
-  const { name, ...userData } = user;
+  const { name, ...userData } = payload;
   const updatedUserData: Partial<IUser> = { ...userData };
+
   if (name && Object.keys(name).length > 0) {
     Object.keys(name).forEach(key => {
       const keyName = `name${key}` as keyof Partial<IUser>;
@@ -42,7 +105,7 @@ const updateUser = async (
     });
   }
 
-  const result = await User.findOneAndUpdate({ _id: id }, updatedUserData, {
+  const result = await User.findOneAndUpdate({ _id }, updatedUserData, {
     new: true,
   });
   return result;
@@ -50,12 +113,19 @@ const updateUser = async (
 
 //------delete a single user service --------------------------------
 const deleteUser = async (id: string): Promise<IUser | null> => {
+  if (!(await isUserAvailable(id))) {
+    throw new ApiError(
+      httpStatus.NOT_FOUND,
+      'User not found in the Database !',
+    );
+  }
+
   const deletedUser = await User.findByIdAndDelete(id);
   return deletedUser;
 };
 
 export const UserService = {
-  createUser,
+  signUp,
   updateUser,
   getAllUsers,
   getSingleUser,
